modified the files according to validation rules and IActionResult

+++ HotelFinder.API/Controllers/HotelsController.cs
@@ -26,9 +26,10 @@ namespace HotelFinder.API.Controllers
         /// </summary>
         /// <returns></returns>
         [HttpGet]
-        public List<Hotel> Get()
+        public IActionResult Get()
         {
-            return _hotelService.GetAllHotels();
+            var hotels = _hotelService.GetAllHotels();
+            return Ok(hotels); //response code:200 ,add data to body
         }
 
         /// <summary>
@@ -37,9 +38,15 @@ namespace HotelFinder.API.Controllers
         /// <param name="id"></param>
         /// <returns></returns>
         [HttpGet("{id}")]
-        public Hotel Get(int id)
+        public IActionResult Get(int id)
         {
-            return _hotelService.GetHotelByID(id);
+            var hotel= _hotelService.GetHotelByID(id);
+            if(hotel!=null)
+            {
+                return Ok(hotel);
+            }
+
+            return NotFound();//404
         }
 
         /// <summary>
@@ -48,9 +55,11 @@ namespace HotelFinder.API.Controllers
         /// <param name="hotel"></param>
         /// <returns></returns>
         [HttpPost]
-        public Hotel Post([FromBody] Hotel hotel)
-        {
-            return _hotelService.CreateHotel(hotel);
+        public IActionResult Post([FromBody] Hotel hotel)
+        {           
+                var createdHotel = _hotelService.CreateHotel(hotel);
+                return CreatedAtAction("Get", new { id = createdHotel.Id }, createdHotel); //201 +data
+
         }
 
         /// <summary>
@@ -59,9 +68,13 @@ namespace HotelFinder.API.Controllers
         /// <param name="hotel"></param>
         /// <returns></returns>
         [HttpPut]
-        public Hotel Put([FromBody] Hotel hotel)
+        public IActionResult Put([FromBody] Hotel hotel)
         {
-            return _hotelService.UpdateHotel(hotel);
+            if(_hotelService.GetHotelByID(hotel.Id)!=null)
+            {
+                return Ok(_hotelService.UpdateHotel(hotel)); //200+data
+            }
+            return NotFound();
 
         }
 
@@ -70,10 +83,15 @@ namespace HotelFinder.API.Controllers
         /// </summary>
         /// <param name="id"></param>
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
-             _hotelService.DeleteHotel(id);
+            if (_hotelService.GetHotelByID(id) != null)
+            {
+                _hotelService.DeleteHotel(id);
+                return Ok(); //200+data
+            }
 
+            return NotFound();
         }
     }
 }
